Tidy up tests after GW relaunch
